feat(level): adding pathfinding service

+++ Assets/Code/NonebNi/Core/Tiles/TileData.cs
@@ -11,10 +11,10 @@ namespace NonebNi.Core.Tiles
     {
         public static readonly TileData Default = new TileData("Default", 1);
 
-        [SerializeField] private float weight;
+        [SerializeField] private int weight;
         [SerializeField] private string name;
 
-        public TileData(string name, float weight)
+        public TileData(string name, int weight)
         {
             this.name = name;
             this.weight = weight;
@@ -22,7 +22,7 @@ namespace NonebNi.Core.Tiles
 
         public string Name => name;
 
-        public float Weight => weight;
+        public int Weight => weight;
 
         public void CopyValueFrom(TileData tileData)
         {
